% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_canonical_transcripts.R
\name{get_canonical_transcripts}
\alias{get_canonical_transcripts}
\title{Return Only Canonical Transcripts}
\usage{
get_canonical_transcripts(snpeff.dt)
}
\arguments{
\item{chr.filter}{Vector containing the chromosomes you want to keep in the
data.frame}
}
\value{
A data.frame of a SnpEff database dump file that only contains the
        canonical transcripts
}
\description{
This function will return a data.frame of a SnpEff database dump file that
only contains the canonical transcripts
}

