% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load.R
\name{load_snpeff_db_dump}
\alias{load_snpeff_db_dump}
\title{Load SnpEff Database Dump Text File}
\usage{
load_snpeff_db_dump(dumpfile)
}
\arguments{
\item{dumpfile}{Text dumpfile from the SnpEff dump command}
}
\value{
A data.table of a SnpEff database dump file
}
\description{
This function will wraps around the datatable::fread() function to allow for
faster and efficient reading of the SnpEff database dump files
}

