% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_TSS_window.R
\name{get_TSS_window}
\alias{get_TSS_window}
\title{Return the interval}
\usage{
get_TSS_window(snpeff.dt, range.len)
}
\arguments{
\item{snpeff.dt}{SnpEff database dump data.table loaded through the
load_snpeff_db_dump() function}

\item{range.len}{Integer value indicating the number of bases to include
       up- and down-stream of the transcription start site}
}
\value{
A data.table of contain the chr, upstream, downstream, and geneName
        of the around the TSS region
}
\description{
This function will return a window of range.len upstream and downstream
of every transcription start site int he snpeff.dt
}

